refactor(observability): 归位 Trace 基础模块
