+++ src/observability/analysis/coverage-analyzer.ts
@@ -19,7 +19,7 @@ import {
 } from 'node:fs';
 import type { ToolCallInfo } from '../../executors/contracts/trace.js';
 import type { AnalysisEntry } from './contracts.js';
-import { toolCallQuery } from '../tool-search.js';
+import { toolCallQuery } from '../trace/tool-search.js';
 import { isToolCallSuccess } from '../../executors/tool-call-status.js';
 
 export interface KnowledgeEntry {
