+++ src/observability/codex-conversation-index.ts
@@ -1,6 +1,6 @@
 import { closeSync, openSync, readSync, statSync } from 'node:fs';
 import type { ExperienceTurnStatus } from './contracts/experience.js';
-import type { TraceToolStatus } from './trace-ir.js';
+import type { TraceToolStatus } from './trace/trace-ir.js';
 import { codexUserDisplayText, codexUserMessageOrigin } from './codex-protocol.js';
 import {
   codexRuntimeToolOutcomeFromPayload,
