+++ src/observability/codex-tool-status.ts
@@ -1,4 +1,4 @@
-import type { TraceToolStatus } from './trace-ir.js';
+import type { TraceToolStatus } from './trace/trace-ir.js';
 import { isToolResultFailureText } from '../executors/tool-call-status.js';
 
 export interface CodexToolOutcome {
