+++ src/observability/skill-health-analyzer.ts
@@ -25,7 +25,7 @@ import {
   type SkillSegment,
 } from './trace-adapter.js';
 import { legacyCcSessionToTraceSession } from './trace-source.js';
-import { createTraceSessionIndex } from './trace-session-index.js';
+import { createTraceSessionIndex } from './trace/session-index.js';
 import { setOwnRecordValue, sumRecordCounts } from '../shared/record-count.js';
 import { checkedSumTokenCounts } from '../shared/token-usage.js';
