+++ src/observability/trace-adapter.ts
@@ -8,7 +8,7 @@
 import type { TraceIngestionSummary } from './contracts/trace.js';
 import type { AnalysisEntry } from './analysis/contracts.js';
 import { loadTraceCorpus } from './trace-source.js';
-import type { TraceSession } from './trace-ir.js';
+import type { TraceSession } from './trace/trace-ir.js';
 import { segmentTraceBySkill, segmentsToAnalysisEntries, type SkillSegment } from './trace-segmenter.js';
 
 export type {
@@ -31,7 +31,7 @@ export type {
   TraceToolResultEvent,
   TraceToolStatus,
   TraceUsageEvent,
-} from './trace-ir.js';
+} from './trace/trace-ir.js';
 export type { TraceIngestionSummary } from './contracts/trace.js';
 export { loadCcSessions, loadTraceCorpus, loadTraceSessions } from './trace-source.js';
 export {
