+++ src/observability/trace-attribution.ts
@@ -1,7 +1,7 @@
 /** Skill attribution rules for trace records. */
 
 import type { CcAssistantRecord, CcUserRecord } from './trace-source.js';
-import type { TraceMessageEvent, TraceToolCallEvent } from './trace-ir.js';
+import type { TraceMessageEvent, TraceToolCallEvent } from './trace/trace-ir.js';
 import { extractCodexExecCommands } from './codex-exec-command.js';
 
 // ---------- Skill signal detection ----------
