+++ src/observability/trace/ingestion.ts
@@ -1,5 +1,5 @@
-import type { TraceIngestionSummary } from './contracts/trace.js';
-import { sumRecordCounts } from '../shared/record-count.js';
+import type { TraceIngestionSummary } from '../contracts/trace.js';
+import { sumRecordCounts } from '../../shared/record-count.js';
 
 export interface TraceIngestionNotice {
   level: 'warning' | 'info';
