+++ src/observability/trace/trace-ir.ts
@@ -1,12 +1,12 @@
 import { createHash } from 'node:crypto';
 import { realpathSync } from 'node:fs';
 import { resolve } from 'node:path';
-import type { ToolCallStatus, ToolCallStatusSource } from '../executors/contracts/trace.js';
-import type { TraceIngestionSummary, TraceSourceKind, TraceSourceMetadata } from './contracts/trace.js';
-import { normalizeRfc3339Timestamp } from '../shared/timestamp.js';
-import type { NormalizedToolIdentity } from '../shared/tool-identity.js';
+import type { ToolCallStatus, ToolCallStatusSource } from '../../executors/contracts/trace.js';
+import type { TraceIngestionSummary, TraceSourceKind, TraceSourceMetadata } from '../contracts/trace.js';
+import { normalizeRfc3339Timestamp } from '../../shared/timestamp.js';
+import type { NormalizedToolIdentity } from '../../shared/tool-identity.js';
 
-export type { TraceSourceKind } from './contracts/trace.js';
+export type { TraceSourceKind } from '../contracts/trace.js';
 export type TraceRole = 'standalone' | 'main' | 'subagent';
 export type TraceMessageRole = 'user' | 'assistant' | 'system';
 export type TraceMessageOrigin = 'human' | 'runtime' | 'skill-context' | 'synthetic';
