+++ test/observability/trace/tool-search.test.ts
@@ -4,8 +4,8 @@ import {
   isFailedSearchToolCall,
   isSearchToolCall,
   toolCallQuery,
-} from '../../src/observability/tool-search.js';
-import type { ToolCallInfo } from '../../src/executors/contracts/trace.js';
+} from '../../../src/observability/trace/tool-search.js';
+import type { ToolCallInfo } from '../../../src/executors/contracts/trace.js';
 
 function bashTc(command: string): ToolCallInfo {
   return {
